final edit before final turn-in
