+++ src/pages/ViewDeck.tsx
@@ -1,5 +1,5 @@
-import { ReactElement, useContext, useState } from "react";
-import { DeckContext } from "../context";
+import { ReactElement, /* useContext, */ useState } from "react";
+// import { DeckContext } from "../context";
 import { useLoaderData } from "react-router-dom";
 import { IDeck } from "../interfaces";
 import { DecklistDisplay } from "../components";
@@ -7,7 +7,7 @@ import { sortDeck } from "../utils";
 // displays decklist sorted by card type, showing various stats and other goodies.
 
 export function ViewDeck(): ReactElement {
-	const deckContext = useContext(DeckContext);
+	// const deckContext = useContext(DeckContext);
 	const [loadedDeck, setloadedDeck] = useState<IDeck>(useLoaderData() as IDeck);
 
 	function filterRealCards(deck: IDeck): IDeck {
