+++ src/utils.tsx
@@ -1,4 +1,4 @@
-import { ICard, IDeck } from "./interfaces";
+import { ICard, IDeck, IDecklistEntry, IDeckStrings } from "./interfaces";
 
 export const baseURL = "https://api.scryfall.com";
 
@@ -467,180 +467,140 @@ export function areAllCardsVerified(inputDeck: IDeck | null): boolean {
 	return soFarSoGood;
 }
 
-export function exampleTransformCard() {
+export function createStringsFromDeck(input: IDeck): IDeckStrings {
+	console.log("in createStringsFromDeck...");
+	console.log("input.ID", input.id);
+	console.log("input.name", input.name);
+	console.log("input.main", input.main);
+	console.log("input.sideboard", input.sideboard);
+
+	// input: an IDeck object.
+	// output: object with 4 strings formatted as intended for the 4 decklist form textboxes.
+	// example input:
+	/*
+		{
+			name: "Hello deck!",
+			main: [
+				{name: "Mox opal", count: 4},
+				{name: "Mox lotus", count: 4}
+			],
+			sideboard: [
+				{name: "Mox opal", count: 4},
+				{name: "Mox lotus", count: 4}
+			]
+		}
+		*/
+	// example output:
+	/* 
+		[
+			"Hello deck!",
+			"4 Mox opal\n4 Mox lotus",
+			"4 Mox opal\n4 Mox lotus",
+		]
+		*/
+
+	const idStrOutput = input.id ? input.id : generateUniqueID();
+	const nameStrOutput = input.name;
+	let mainStrOutput = "";
+	let sideboardStrOutput = "";
+
+	for (const entry of input.main) {
+		mainStrOutput = `${mainStrOutput}\n${entry.count} ${entry.name}`;
+	}
+
+	for (const entry of input.sideboard) {
+		sideboardStrOutput = `${sideboardStrOutput}\n${entry.count} ${entry.name}`;
+	}
+
+	console.log("idStrOutput trim:", idStrOutput);
+	console.log("nameStrOutput trim:", nameStrOutput);
+	console.log("mainStrOutput trim:", mainStrOutput);
+	console.log("sideboardStrOutput trim:", sideboardStrOutput);
+
 	return {
-		name: "Arguel's Blood Fast // Temple of Aclazotz",
-		count: 1,
-		card_info: {
-			object: "card",
-			id: "c4ac7570-e74e-4081-ac53-cf41e695b7eb",
-			oracle_id: "be2a4bc4-8af6-48c5-9421-32d26272e71a",
-			multiverse_ids: [435243, 435244],
-			mtgo_id: 65194,
-			mtgo_foil_id: 65195,
-			arena_id: 66143,
-			tcgplayer_id: 145261,
-			cardmarket_id: 301387,
-			name: "Arguel's Blood Fast // Temple of Aclazotz",
-			lang: "en",
-			released_at: "2017-09-29",
-			uri: "https://api.scryfall.com/cards/c4ac7570-e74e-4081-ac53-cf41e695b7eb",
-			scryfall_uri: "https://scryfall.com/card/xln/90/arguels-blood-fast-temple-of-aclazotz?utm_source=api",
-			layout: "transform",
-			highres_image: true,
-			image_status: "highres_scan",
-			cmc: 2,
-			type_line: "Legendary Enchantment // Legendary Land",
-			color_identity: ["B"],
-			keywords: ["Transform"],
-			produced_mana: ["B"],
-			card_faces: [
-				{
-					object: "card_face",
-					name: "Arguel's Blood Fast",
-					mana_cost: "{1}{B}",
-					type_line: "Legendary Enchantment",
-					oracle_text: "{1}{B}, Pay 2 life: Draw a card.\nAt the beginning of your upkeep, if you have 5 or less life, you may transform Arguel's Blood Fast.",
-					colors: ["B"],
-					flavor_text: "Arguel's vision led him into the jungles of Ixalan . . .",
-					artist: "Daarken",
-					artist_id: "e607a0d4-fc12-4c01-9e3f-501f5269b9cb",
-					illustration_id: "a1e9f54d-adde-4c8a-acd0-5c63c2fece55",
-					image_uris: {
-						small: "https://cards.scryfall.io/small/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						normal: "https://cards.scryfall.io/normal/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						large: "https://cards.scryfall.io/large/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						png: "https://cards.scryfall.io/png/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.png?1562563598",
-						art_crop: "https://cards.scryfall.io/art_crop/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						border_crop: "https://cards.scryfall.io/border_crop/front/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-					},
-				},
-				{
-					object: "card_face",
-					name: "Temple of Aclazotz",
-					mana_cost: "",
-					type_line: "Legendary Land",
-					oracle_text: "(Transforms from Arguel's Blood Fast.)\n{T}: Add {B}.\n{T}, Sacrifice a creature: You gain life equal to the sacrificed creature's toughness.",
-					colors: [],
-					flavor_text: ". . . to the lost temple of a bat-god of night, eternal enemy to the Threefold Sun.",
-					artist: "Daarken",
-					artist_id: "e607a0d4-fc12-4c01-9e3f-501f5269b9cb",
-					illustration_id: "872bd01d-10b1-4050-84e8-bf128e624b6d",
-					image_uris: {
-						small: "https://cards.scryfall.io/small/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						normal: "https://cards.scryfall.io/normal/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						large: "https://cards.scryfall.io/large/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						png: "https://cards.scryfall.io/png/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.png?1562563598",
-						art_crop: "https://cards.scryfall.io/art_crop/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-						border_crop: "https://cards.scryfall.io/border_crop/back/c/4/c4ac7570-e74e-4081-ac53-cf41e695b7eb.jpg?1562563598",
-					},
-				},
-			],
-			all_parts: [
-				{
-					object: "related_card",
-					id: "4a8b3030-5495-404b-938c-8c4a387a1341",
-					component: "combo_piece",
-					name: "Ixalan Checklist",
-					type_line: "Card",
-					uri: "https://api.scryfall.com/cards/4a8b3030-5495-404b-938c-8c4a387a1341",
-				},
-				{
-					object: "related_card",
-					id: "c4ac7570-e74e-4081-ac53-cf41e695b7eb",
-					component: "combo_piece",
-					name: "Arguel's Blood Fast // Temple of Aclazotz",
-					type_line: "Legendary Enchantment // Legendary Land",
-					uri: "https://api.scryfall.com/cards/c4ac7570-e74e-4081-ac53-cf41e695b7eb",
-				},
-				{
-					object: "related_card",
-					id: "31d2e591-6df4-4445-a16c-c7a576ff3e20",
-					component: "combo_piece",
-					name: "High Marshal Arguel",
-					type_line: "Legendary Creature — Vampire Knight",
-					uri: "https://api.scryfall.com/cards/31d2e591-6df4-4445-a16c-c7a576ff3e20",
-				},
-			],
-			legalities: {
-				standard: "not_legal",
-				future: "not_legal",
-				historic: "legal",
-				timeless: "legal",
-				gladiator: "legal",
-				pioneer: "legal",
-				explorer: "legal",
-				modern: "legal",
-				legacy: "legal",
-				pauper: "not_legal",
-				vintage: "legal",
-				penny: "legal",
-				commander: "legal",
-				oathbreaker: "legal",
-				standardbrawl: "not_legal",
-				brawl: "legal",
-				alchemy: "not_legal",
-				paupercommander: "not_legal",
-				duel: "legal",
-				oldschool: "not_legal",
-				premodern: "not_legal",
-				predh: "not_legal",
-			},
-			games: ["arena", "paper", "mtgo"],
-			reserved: false,
-			foil: true,
-			nonfoil: true,
-			finishes: ["nonfoil", "foil"],
-			oversized: false,
-			promo: false,
-			reprint: false,
-			variation: false,
-			set_id: "fe0dad85-54bc-4151-9200-d68da84dd0f2",
-			set: "xln",
-			set_name: "Ixalan",
-			set_type: "expansion",
-			set_uri: "https://api.scryfall.com/sets/fe0dad85-54bc-4151-9200-d68da84dd0f2",
-			set_search_uri: "https://api.scryfall.com/cards/search?order=set&q=e%3Axln&unique=prints",
-			scryfall_set_uri: "https://scryfall.com/sets/xln?utm_source=api",
-			rulings_uri: "https://api.scryfall.com/cards/c4ac7570-e74e-4081-ac53-cf41e695b7eb/rulings",
-			prints_search_uri: "https://api.scryfall.com/cards/search?order=released&q=oracleid%3Abe2a4bc4-8af6-48c5-9421-32d26272e71a&unique=prints",
-			collector_number: "90",
-			digital: false,
-			rarity: "rare",
-			artist: "Daarken",
-			artist_ids: ["e607a0d4-fc12-4c01-9e3f-501f5269b9cb"],
-			border_color: "black",
-			frame: "2015",
-			frame_effects: ["compasslanddfc"],
-			security_stamp: "oval",
-			full_art: false,
-			textless: false,
-			booster: true,
-			story_spotlight: false,
-			edhrec_rank: 6205,
-			penny_rank: 2764,
-			prices: {
-				usd: "0.26",
-				usd_foil: "0.67",
-				usd_etched: null,
-				eur: "0.33",
-				eur_foil: "0.75",
-				tix: "0.02",
-			},
-			related_uris: {
-				gatherer: "https://gatherer.wizards.com/Pages/Card/Details.aspx?multiverseid=435243&printed=false",
-				tcgplayer_infinite_articles:
-					"https://tcgplayer.pxf.io/c/4931599/1830156/21018?subId1=api&trafcat=infinite&u=https%3A%2F%2Finfinite.tcgplayer.com%2Fsearch%3FcontentMode%3Darticle%26game%3Dmagic%26partner%3Dscryfall%26q%3DArguel%2527s%2BBlood%2BFast%2B%252F%252F%2BTemple%2Bof%2BAclazotz",
-				tcgplayer_infinite_decks:
-					"https://tcgplayer.pxf.io/c/4931599/1830156/21018?subId1=api&trafcat=infinite&u=https%3A%2F%2Finfinite.tcgplayer.com%2Fsearch%3FcontentMode%3Ddeck%26game%3Dmagic%26partner%3Dscryfall%26q%3DArguel%2527s%2BBlood%2BFast%2B%252F%252F%2BTemple%2Bof%2BAclazotz",
-				edhrec: "https://edhrec.com/route/?cc=Arguel%27s+Blood+Fast",
-			},
-			purchase_uris: {
-				tcgplayer: "https://tcgplayer.pxf.io/c/4931599/1830156/21018?subId1=api&u=https%3A%2F%2Fwww.tcgplayer.com%2Fproduct%2F145261%3Fpage%3D1",
-				cardmarket: "https://www.cardmarket.com/en/Magic/Products/Singles/Ixalan/Arguels-Blood-Fast-Temple-of-Aclazotz?referrer=scryfall&utm_campaign=card_prices&utm_medium=text&utm_source=scryfall",
-				cardhoarder: "https://www.cardhoarder.com/cards/65194?affiliate_id=scryfall&ref=card-profile&utm_campaign=affiliate&utm_medium=card&utm_source=scryfall",
-			},
-		},
-		is_real: true,
+		idStr: idStrOutput.trim(),
+		nameStr: nameStrOutput.trim(),
+		mainStr: mainStrOutput.trim(),
+		sideboardStr: sideboardStrOutput.trim(),
 	};
+}
+
+export function createDeckFromStrings(inputID: string, inputName: string, inputMain: string, inputSB: string): IDeck {
+	// 241014 TODO
+	// 1. turn strings into arrays of lines
+	// 2. for main: turn each line into decklistEntry
+	// 3. for sb: turn each line into decklistEntry
+	// 4. combine name, main, sb into IDeck object.
+	// 5. return IDeck object
+
+	if (!inputID) {
+		inputID = generateUniqueID();
+	}
+
+	// 1. turn strings into arrays of lines
+	let linesMain: string[] = inputMain.split("\n");
+	let linesSB: string[] = inputSB.split("\n");
+
+	// 2. for main: turn each line into decklistEntry
+	let deckMain: IDecklistEntry[] = [];
+	for (const line of linesMain) {
+		let entry: IDecklistEntry | null = toDecklistEntry(line);
+		console.log("Maindeck entry:", entry);
+		//add entry to list if entry is a valid decklistEntry (not null)
+		if (entry !== null) {
+			deckMain.push(entry);
+		}
+	}
+
+	// 3. for sb: turn each line into decklistEntry
+	let deckSB: IDecklistEntry[] = [];
+	for (const line of linesSB) {
+		let entry: IDecklistEntry | null = toDecklistEntry(line);
+		console.log("SB entry:", entry);
+		//add entry to list if entry is a valid decklistEntry (not null)
+		if (entry !== null) {
+			deckSB.push(entry);
+		}
+	}
+
+	console.log("deck ID: ", inputID);
+	console.log("deck Name: ", inputName);
+	console.log("deck Main: ", deckMain);
+	console.log("deck SB: ", deckSB);
+
+	return {
+		id: inputID,
+		name: inputName,
+		main: deckMain,
+		sideboard: deckSB,
+	};
+}
+
+export function toDecklistEntry(input: string): IDecklistEntry | null {
+	// example input: "4 mox opal"
+	// exmple output: {name: "mox opal", count: 4}
+
+	if (input.length === 0) {
+		//data validation
+		return null;
+	}
+
+	// ...extract name and count:
+	const cardCountStr = input.split(" ", 1)[0]; // '4 mox opal' -> '4'
+	const cardNameStr = input.slice(cardCountStr.length + 1); // '4 mox opal' -> 'mox opal'
+	const decklistEntry = {
+		name: cardNameStr,
+		count: parseInt(cardCountStr),
+	};
+	// console.log("decklistEntry", decklistEntry);
+
+	if (
+		// more data validation (for culling invalid decklist entries)
+		decklistEntry.name.length == 0 ||
+		decklistEntry.count == 0 ||
+		Number.isNaN(decklistEntry.count)
+	) {
+		return null;
+	}
+	return decklistEntry; // format: {name: "Mox Opal", count: 4}
 }
